Fix MB-845
This fixes the issue "Unable to delete topics from UI when a subscriber crash
without closing connection"  reported in MB-845.

URL: https://wso2.org/jira/browse/MB-845

When a client get killed before closing the AMQP connection properly, Mina transport
handler detects this and initiate closing the AMQP connection belonging to that user.
But it never set the "AuthorizedSubject" for the current thread. Since no user is
associated with the current thread, updating registry for subscription deletion fails.

In the fix I set the AuthorizedSubject in the "closed" method of the ProtocolEngine.