Fix for issue https://wso2.org/jira/browse/MB-890
When an exception occur on ConcurrentBatchProcessor while calling
onEvent(...) of relevant even handler exception is handle without
clearing the current batched event list. Therefore after handling
the exception, in the next iteration batch processor feeds the
event handler with a list with new events including the events from
erronous previous iteration.
This leads to feeding events to handler that are already cleared.

Fixed this by clearing the eventlist when an exception is caught.