Slots are not returned when disconnecting subscribers that have failed over from one MB node.
This happens because the internal map "trackedSubscriptions" in OrphanedSlotHandler
does not have the subscribedNode as part of the key. This causes the newly failed
over subscription on MB2 node to be deleted from this map, which in return skips
the slot return logic when this subscription disconnects.

Note: Fix was done by @hastef88