MB-1892 issue fixed.
Null subscription assigned if the subscriber connection is not a local. But when rendering nondurable subscriptions, we are fetching all subscriptions. Therefore, the condition added to check whether the subscription is local or not and based on that protocol queue name will retrieve. Also, NullSubscription class modified to throw UnsupportedOperationException.

Public JIRA - https://wso2.org/jira/browse/MB-1892