MB-1911 issue fixed
Durable topic subscription search ends up in a classcast exception when subscription added to the same
topic from multiple nodes. In the topic subscription filtering logic cast AndesSubscription to
DurableTopicSubscriber. But, we are getting an AndesSubscription object for the subscription which is
on the remote node. Also, filtering logic not work if there are any leading or trailing whitespaces.
Therefore, trim() the input string.