Add YAML based configuration support.
Add carbon-config dependancy and the plugin for config doc generation.
Add Configuration service singleton to fetch the root config instance.
Add the broker config root model and a model for subset of coordination
configs as a sample.
Use BrokerConfigurationService.getInstance().getBrokerConfiguration() to
fetch a config value.
NOTE : This commit only add support for yaml based configs, conversion
of broker.xml to deployment.yaml is not done.